% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_helpers.R
\name{col_num_successes}
\alias{col_num_successes}
\title{Count Successes in a \code{list} Column}
\usage{
col_num_successes(df, col, .p, name = .successes)
}
\arguments{
\item{df}{A \code{tibble} containing the outcome column.}

\item{col}{A \code{list} column of outcomes to inspect.}

\item{.p}{An atomic object to match or a Boolean predicate function.}

\item{name}{Name of new column (default = .successes)}
}
\value{
A \code{tibble} with all original columns and a new column of success counts.
}
\description{
Create a new \code{integer} column contain the number of successes in corresponding entry of a list column of outcomes.
}
\examples{
library(tidyr)
library(purrr)
library(dplyr)

df <- tibble(lengths = 1:5) \%>\%
   mutate(.outcome = map(lengths, \(n) rep_len('H', n))) \%>\%
   walk(glimpse)

df \%>\%
 col_num_successes(.outcome, 'H') \%>\%
 glimpse
}
