% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{col_sample}
\alias{col_sample}
\title{Sample from a List Column}
\usage{
col_sample(.df, .col, size, prob = NULL, replace = FALSE, .as = .outcome)
}
\arguments{
\item{.df}{A tibble containing the sample space in a column and one row per trial.}

\item{.col}{The list column to be sampled.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled.}

\item{replace}{should sampling be with replacement?}

\item{.as}{the name of the resulting outcome column (default = .outcome)}
}
\value{
An updated tibble containing the outcomes in a list column
}
\description{
Take a sample from a list column \code{.col} from the tibble \code{.df}
}
\examples{
library(dplyr)

# Flip a fair coin 3 times (10 trials)
tribble(~coin,
        c('H', 'T')) \%>\%
  add_trials(10) \%>\%
  col_sample(coin, 3, replace = TRUE)

# Same experiment, but rename the outcome column
tribble(~coin,
        c('H', 'T')) \%>\%
   add_trials(10) \%>\%
   col_sample(coin, 3, replace = TRUE, .as = three.tosses)
}
