% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_prob}
\alias{estimate_prob}
\title{Estimate probabilities}
\usage{
estimate_prob(df, col)
}
\arguments{
\item{df}{A \code{data.frame}/\code{tibble}}

\item{col}{A Boolean/logical column}
}
\value{
A table containing the estimated probability.
}
\description{
Estimate probabilities for a Boolean/logical vector \code{x} by using
\itemize{
\item \code{df \%>\% estimate_prob(x)} estimate the marginal probability estimate of \code{x==TRUE}
\item \code{df \%>\% group_by(y) \%>\% estimate_prob(x)} to estimate the conditional probabilities of \code{x=TRUE} given \code{y}
}
}
\examples{
library(dplyr)
library(purrr)

coin <- c('H', 'T')
prob <- c(0.55, 0.45)

#Example 1.  Estimate P(First is H)
init_trials(1000) \%>\%
   mutate(three.tosses = map(.trial, \(i) sample(coin, 3, replace=TRUE, prob = prob)),
          first.head = map_lgl(three.tosses, \(x) pluck(x, 1) == 'H'),
   ) -> df

df \%>\% glimpse # Or use `View(df)` to see the whole tableau

df \%>\%
  estimate_prob(first.head)

# Package the whole experiment in one pipe

(init_trials(1000)
   \%>\% mutate(three.tosses = map(.trial, \(i) sample(coin, 3, replace=TRUE, prob = prob)),
          first.head = map_lgl(three.tosses, \(x) pluck(x, 1) == 'H'),
          )
   # \%>\% walk(glimpse) #uncomment to inspect first few rows of tableau
   \%>\% estimate_prob(first.head)
 )

 # Example 2. Estimate P(first is H | second is T)
(init_trials(1000)
   \%>\% mutate(three.tosses = map(.trial, \(i) sample(coin, 3, replace=TRUE, prob = prob)),
              first.head = map_lgl(three.tosses, \(x) pluck(x, 1) == 'H'),
              second.tail = map_lgl(three.tosses, \(x) pluck(x, 2) == 'T'),
              )
   # \%>\% walk(glimpse) #uncomment to inspect first few rows of tableau
   \%>\% group_by(second.tail)
   \%>\% estimate_prob(first.head)
)
}
