% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_helpers.R
\name{get_label_count}
\alias{get_label_count}
\title{Counting Labels}
\usage{
get_label_count(.data, .col, .label)
}
\arguments{
\item{.data}{A data frame/tibble.}

\item{.col}{A list column of character variables.}

\item{.label}{The label being counted.}
}
\value{
The original .data with a new count column
}
\description{
Counts the number of occurrences of a given label in a list column of
of character vectors, storing the result in a new column named "num_{.label}".
This method was designed for counting the outcomes in a multinomial setting.
}
\examples{
grades <- c('A', 'B', 'C', 'D', 'F')
probs <- c(0.23, 0.26, 0.21, 0.18, 0.12)
n <- 5

replicate(5, sample(grades, n, prob = probs, replace = TRUE), .as = raw.grades) \%>\%
  get_label_count(raw.grades, 'A')
}
\seealso{
\code{\link[=multicounts]{multicounts()}}
}
