% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{init_trials}
\alias{init_trials}
\title{Initial a Probability Tableau}
\usage{
init_trials(n)
}
\arguments{
\item{n}{The number of trials.}
}
\value{
A \code{tibble} with \code{n} rows and a \code{.trial = 1:n} column.
}
\description{
Creates a \code{tibble} with \code{n} rows and a \code{.trial = 1:n} column
}
\examples{
library(dplyr)
library(purrr)

init_trials(10) # Blank tableau

coin <- c('H', 'T')
init_trials(10) \%>\% mutate(.outcome = map(.trial, \(i) sample(coin, 5, replace = TRUE)))
}
