% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictate.R
\name{is.equal}
\alias{is.equal}
\alias{not.equal}
\alias{strictly.greater}
\alias{at.least}
\alias{strictly.less}
\alias{at.most}
\title{Predicate functions}
\usage{
is.equal(x, y)

not.equal()

strictly.greater(x, y)

at.least(x, y)

strictly.less(x, y)

at.most(x, y)
}
\arguments{
\item{x}{atomic vectors, symbols, calls, or other objects for which methods have}

\item{y}{been written}
}
\value{
Boolean
}
\description{
Predicate comparison functions in plain English.
}
\examples{
library(dplyr)
x <- stats::rnorm(20)
x \%>\% strictly.less(1)
x[x \%>\% strictly.less(1)]
}
