% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labellers.R
\name{label_value_with}
\alias{label_value_with}
\alias{label_both_with}
\alias{label_context_with}
\alias{label_parsed_with}
\alias{label_wrap_gen_with}
\title{Labeller helper functions}
\usage{
label_value_with(multi_line = TRUE)

label_both_with(multi_line = TRUE, sep = ": ")

label_context_with(multi_line = TRUE, sep = ": ")

label_parsed_with(multi_line = TRUE)

label_wrap_gen_with(width = 25, multi_line = TRUE)
}
\arguments{
\item{labels}{Data frame of labels. Usually contains only one element, but faceting over multiple factors entails multiple label variables.}

\item{Whether}{to display the labels of multiple factors on separate lines.}

\item{String}{separating variables and values.}

\item{Maximum}{number of characters before wrapping the strip.}
}
\value{
a labeller function
}
\description{
Returns a labeller function with the provided arguements, making it easier
to provide a custom labeller.
}
\examples{
library(ggplot2)

mtcars$cyl2 <- factor(mtcars$cyl, labels = c("alpha", "beta", "gamma"))
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

# Displaying both the values and the variables with default `sep`.
p + facet_grid(. ~ cyl, labeller = label_both)

# Displaying both the values and the variables with default `sep`.
p + facet_grid(. ~ cyl, labeller = label_both_with(sep = "="))
}
