% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_helpers.R
\name{multicounts}
\alias{multicounts}
\title{Count Multiple Labels}
\usage{
multicounts(.data, .col, .labels)
}
\arguments{
\item{.data}{A data frame/tibble.}

\item{.col}{A list column of character variables.}

\item{.labels}{A vector of labels to be counted.}
}
\value{
The original .data with a new count columns, one for each entry in .labels
}
\description{
Counts the number of occurrences of a given collection of labels in a list column of
of character vectors, storing the result in a new columns named "num_{label}".
This method was designed for counting the outcomes in a multinomial setting.
}
\examples{
grades <- c('A', 'B', 'C', 'D', 'F')
probs <- c(0.23, 0.26, 0.21, 0.18, 0.12)
n <- 5

replicate(5, sample(grades, n, prob = probs, replace = TRUE), .as = raw.grades) \%>\%
  multicounts(raw.grades, grades)
}
\seealso{
\code{\link[=get_label_count]{get_label_count()}}
}
