% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_helper.R
\name{num.successes}
\alias{num.successes}
\title{Count the number of successes}
\usage{
num.successes(col, p)
}
\arguments{
\item{col}{A list of outcomes.}

\item{p}{Either an atomic element or predicate functions.}
}
\value{
The number of successes.
}
\description{
Count the number of successes
}
\examples{
library(dplyr)
x <- list(c('Red', 'Green', 'Blue'))
x \%>\% num.successes('Red')
x \%>\% num.successes(\(x) x != 'Green')
}
