% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{parameters}
\alias{parameters}
\title{Row-wise parameter space creation}
\usage{
parameters(...)
}
\arguments{
\item{...}{Arguments specifying the structure of a tibble.
Variable names should be formulas, and may only appear before the data.
These arguments are processed with rlang::list2() and support unquote via
!! and unquote-splice via !!!.}
}
\value{
A tibble
}
\description{
Create tibbles of unstacked parameters using an easier to read row-by-row
layout. This is useful for creating an initial collection of parameters when
simulating a family of parametric distributions. Please see tibble-package
and \code{tribble} function for a general introduction.
}
\examples{
parameters(
   ~n,      ~p,
   c(5,10), c(0.4, 0.6))
}
