% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_table}
\alias{sample_table}
\title{Random Samples of a Dataframe}
\usage{
sample_table(.df, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{.df}{A \code{data.frame}/\code{tibble}}

\item{size}{A positive number representing the number of rows to be sampled.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled.  Can either be a column of weights contained in \code{.df.} or a separate vector with a length of \code{nrow(.df)}}
}
\value{
A \code{data.frame}/\code{tibble}
}
\description{
\code{sample_table} samples the rows of \code{.df} returning the resulting data frame
}
\examples{
library(purrr)

biased_coin <- data.frame(toss = c(rep('H', 55), rep('T', 45)))

sample_table(biased_coin, 5, replace = TRUE)

map(1:10, \(i) sample_table(biased_coin, 5, replace = TRUE))


urn <- data.frame(draw = c(rep('R', 4), rep('W', 3)))
urn

sample_table(urn, 3, replace = FALSE)

coin <- data.frame(toss = c('H', 'T'))
ps <- c(0.55, 0.45)
sample_table(coin, 3, replace = TRUE, prob = ps)

coin_w_weights <- data.frame(toss = c('H', 'T'), weight = c(0.55, 0.45))
coin_w_weights

sample_table(coin_w_weights, 3, replace = TRUE, prob = weight)
}
