% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{tabulate}
\alias{tabulate}
\title{Construct tables of probability estimates}
\usage{
tabulate(df, col, prefix = "X = ")
}
\arguments{
\item{df}{A \code{tibble} containing one or more variables.}

\item{col}{The column of interest}

\item{prefix}{The string used as a prefix in the output.}
}
\value{
A table of (row-wise) relative frequencies.
}
\description{
Estimate probabilities  with \code{tabulate} by using
\itemize{
\item \code{df \%>\% tabulate(x)} estimate the marginal distribution of \code{x}
\item \code{df \%>\% group_by(y) \%>\% tabulate(x)} to estimate the conditional distributions of \code{x} given \code{y}
}
}
\examples{
library(dplyr)
library(purrr)

coin <- c('H', 'T')
prob <- c(0.55, 0.45)

# Getting estimated the marginal distribution
init_trials(1000) \%>\%
  mutate(one.toss = map(.trial, \(i) sample(coin, 1, prob = prob))) \%>\%
  tabulate(one.toss)

# Estimating the conditional distributions
init_trials(1000) \%>\%
  mutate(three.tosses = map(.trial, \(i) sample(coin, 3, replace=TRUE, prob = prob)),
         first.toss = map_chr(three.tosses, \(x) x \%>\% pluck(1)),
         second.toss = map_chr(three.tosses, \(x) x \%>\% pluck(2)),
        ) \%>\%
  group_by(first.toss) \%>\%
  tabulate(second.toss, prefix = "second.toss=")
}
