% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{take.one.sample}
\alias{take.one.sample}
\title{Random Samples and Permutations}
\usage{
take.one.sample(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose,
    or a positive integer.  See \sQuote{Details.}}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements
    of the vector being sampled.}
}
\value{
A list of outcomes
}
\description{
Random Samples and Permutations
}
\examples{
take.one.sample(1:10, 3)
}
