% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{add_trials}
\alias{add_trials}
\title{Title}
\usage{
add_trials(df, n)
}
\arguments{
\item{df}{A \code{data.frame}/\code{tibble} containing the experiment's parameters.}

\item{n}{The number of trials.}
}
\value{
A \code{tibble} with the parameter space replicated \code{n} times and a \code{.trial=1:n} column.
}
\description{
Title
}
\examples{
library(tidyr)
tribble(
  ~coin, ~n,
  c('H', 'T'), 5,
  ) \%>\%
  add_trials(10)
}
